refactor(features): extract UI components and improve feature dialog structure
- Move WebFeature class to separate file with static defaults
- Extract IconBrowserDialog to separate component
- Extract FeatureCard builder methods for toggle and link cards
- Add Utils.getString for localized text handling
- Simplify FeatureSettingDialog by using extracted components
- Remove boolean parameter from show() calls
- Improve code organization and maintainability
